winterfell

Module matrix

Two-dimensional data structures used to represent polynomials and polynomial evaluations.

Structs

ColMatrix

A two-dimensional matrix of field elements arranged in column-major order.

ColumnIter

Iterator over columns of ColMatrix.

RowMatrix

A two-dimensional matrix of field elements arranged in row-major order.

Segment

A set of columns of a matrix stored in row-major form.

Functions

build_segments

Returns matrix segments constructed by evaluating polynomials in the specified matrix over the domain defined by twiddles and offsets.

get_evaluation_offsets

Returns a vector of offsets for an evaluation defined by the specified polynomial size, blowup factor and domain offset.