winterfell::matrix

Function build_segments

pub fn build_segments<E, const N: usize>(
    polys: &ColMatrix<E>,
    twiddles: &[<E as FieldElement>::BaseField],
    offsets: &[<E as FieldElement>::BaseField],
) -> Vec<Segment<<E as FieldElement>::BaseField, N>>
where
    E: FieldElement,

Returns matrix segments constructed by evaluating polynomials in the specified matrix over the domain defined by twiddles and offsets.