winterfell::matrix

Struct RowMatrix

pub struct RowMatrix<E>
where
    E: FieldElement,
{ /* private fields */ }

A two-dimensional matrix of field elements arranged in row-major order.

The matrix is represented as a single vector of base field elements for the field defined by E type parameter. The first row_width base field elements represent the first row of the matrix, the next row_width base field elements represent the second row, and so on.

When rows are returned via the RowMatrix::row() method, base field elements are grouped together as appropriate to form elements in E.

In some cases, rows may be padded with extra elements. The number of elements which are accessible via the RowMatrix::row() method is specified by the elements_per_row member.

Implementations

impl<E> RowMatrix<E>
where E: FieldElement,

pub fn evaluate_polys<const N: usize>( polys: &ColMatrix<E>, blowup_factor: usize, ) -> RowMatrix<E>

Returns a new RowMatrix constructed by evaluating the provided polynomials over the domain defined by the specified blowup factor.

The provided polys matrix is assumed to contain polynomials in coefficient form (one polynomial per column). Columns in the returned matrix will contain evaluations of the corresponding polynomials over the domain defined by polynomial size (i.e., number of rows in the polys matrix) and the blowup_factor.

To improve performance, polynomials are evaluated in batches specified by the N type parameter. Minimum batch size is 1.

pub fn evaluate_polys_over<const N: usize>( polys: &ColMatrix<E>, domain: &StarkDomain<<E as FieldElement>::BaseField>, ) -> RowMatrix<E>

Returns a new RowMatrix constructed by evaluating the provided polynomials over the specified StarkDomain.

The provided polys matrix is assumed to contain polynomials in coefficient form (one polynomial per column). Columns in the returned matrix will contain evaluations of the corresponding polynomials over the LDE domain defined by the provided StarkDomain.

To improve performance, polynomials are evaluated in batches specified by the N type parameter. Minimum batch size is 1.

pub fn from_segments<const N: usize>( segments: Vec<Segment<<E as FieldElement>::BaseField, N>>, elements_per_row: usize, ) -> RowMatrix<E>

Returns a new RowMatrix instantiated from the specified matrix segments.

elements_per_row specifies how many base field elements are considered to form a single row in the matrix.

Panics

Panics if

segments is an empty vector.

elements_per_row is greater than the row width implied by the number of segments and N type parameter.

pub fn num_cols(&self) -> usize

Returns the number of columns in this matrix.

pub fn num_rows(&self) -> usize

Returns the number of rows in this matrix.

pub fn get(&self, col_idx: usize, row_idx: usize) -> E

Returns the element located at the specified column and row indexes in this matrix.

Panics

Panics if either col_idx or row_idx are out of bounds for this matrix.

pub fn row(&self, row_idx: usize) -> &[E]

Returns a reference to a row at the specified index in this matrix.

Panics

Panics if the specified row index is out of bounds.

pub fn data(&self) -> &[<E as FieldElement>::BaseField]

Returns the data in this matrix as a slice of field elements.

pub fn commit_to_rows<H, V>(&self, partition_options: PartitionOptions) -> V
where H: ElementHasher<BaseField = <E as FieldElement>::BaseField>, V: VectorCommitment<H>,

Returns a commitment to this matrix.

The commitment is built as follows:

Each row of the matrix is hashed into a single digest of the specified hash function. The result is a vector of digests of length equal to the number of matrix rows.

A vector commitment is computed for the resulting vector using the specified vector commitment scheme.

The resulting vector commitment is returned as the commitment to the entire matrix.