winterfell

Struct DefaultTraceLde

pub struct DefaultTraceLde<E, H, V>
where
    E: FieldElement,
    H: ElementHasher<BaseField = <E as FieldElement>::BaseField>,
    V: VectorCommitment<H>,
{ /* private fields */ }

Contains all segments of the extended execution trace, the commitments to these segments, the LDE blowup factor, and the TraceInfo.

Segments are stored in two groups:

Main segment: this is the first trace segment generated by the prover. Values in this segment will always be elements in the base field (even when an extension field is used).

Auxiliary segments: a list of 0 or more segments for traces generated after the prover commits to the first trace segment. Currently, at most 1 auxiliary segment is possible.

Implementations

impl<E, H, V> DefaultTraceLde<E, H, V>
where E: FieldElement, H: ElementHasher<BaseField = <E as FieldElement>::BaseField>, V: VectorCommitment<H>,

pub fn new( trace_info: &TraceInfo, main_trace: &ColMatrix<<E as FieldElement>::BaseField>, domain: &StarkDomain<<E as FieldElement>::BaseField>, partition_options: PartitionOptions, ) -> (DefaultTraceLde<E, H, V>, TracePolyTable<E>)

Takes the main trace segment columns as input, interpolates them into polynomials in coefficient form, evaluates the polynomials over the LDE domain, commits to the polynomial evaluations, and creates a new DefaultTraceLde with the LDE of the main trace segment and the commitment.

Returns a tuple containing a TracePolyTable with the trace polynomials for the main trace segment and the new DefaultTraceLde.