winterfell

Trait ConstraintEvaluator

pub trait ConstraintEvaluator<E>
where
    E: FieldElement,
{
    type Air: Air<BaseField = <E as FieldElement>::BaseField>;

    // Required method
    fn evaluate<T>(
        self,
        trace: &T,
        domain: &StarkDomain<<E as FieldElement>::BaseField>,
    ) -> CompositionPolyTrace<E>
       where T: TraceLde<E>;
}

Contains logic for evaluating AIR constraints over an extended execution trace.

The logic for evaluating AIR constraints over a single evaluation frame is defined by the Air associated type, and the purpose of this trait is to execute this logic over all evaluation frames in an extended execution trace.

Required Associated Types

type Air: Air<BaseField = <E as FieldElement>::BaseField>

AIR constraints for the computation described by this evaluator.

Required Methods

fn evaluate<T>( self, trace: &T, domain: &StarkDomain<<E as FieldElement>::BaseField>, ) -> CompositionPolyTrace<E>
where T: TraceLde<E>,

Evaluates constraints against the provided extended execution trace, combines them into evaluations of a single polynomial, and returns these evaluations.

Constraints are evaluated over a constraint evaluation domain. This is an optimization because constraint evaluation domain can be many times smaller than the full LDE domain.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.