winterfell

Trait TraceLde

pub trait TraceLde<E>: Sync
where
    E: FieldElement,
{
    type HashFn: ElementHasher<BaseField = <E as FieldElement>::BaseField>;
    type VC: VectorCommitment<Self::HashFn>;

    // Required methods
    fn get_main_trace_commitment(&self) -> <Self::HashFn as Hasher>::Digest;
    fn set_aux_trace(
        &mut self,
        aux_trace: &ColMatrix<E>,
        domain: &StarkDomain<<E as FieldElement>::BaseField>,
    ) -> (ColMatrix<E>, <Self::HashFn as Hasher>::Digest);
    fn read_main_trace_frame_into(
        &self,
        lde_step: usize,
        frame: &mut EvaluationFrame<<E as FieldElement>::BaseField>,
    );
    fn read_aux_trace_frame_into(
        &self,
        lde_step: usize,
        frame: &mut EvaluationFrame<E>,
    );
    fn read_lagrange_kernel_frame_into(
        &self,
        lde_step: usize,
        col_idx: usize,
        frame: &mut LagrangeKernelEvaluationFrame<E>,
    );
    fn query(&self, positions: &[usize]) -> Vec<Queries>;
    fn trace_len(&self) -> usize;
    fn blowup(&self) -> usize;
    fn trace_info(&self) -> &TraceInfo;
}

Contains all segments of the extended execution trace and their commitments.

Segments are stored in two groups:

Main segment: this is the first trace segment generated by the prover. Values in this segment will always be elements in the base field (even when an extension field is used).

Auxiliary segments: a list of 0 or more segments for traces generated after the prover commits to the first trace segment. Currently, at most 1 auxiliary segment is possible.

Required Associated Types

type HashFn: ElementHasher<BaseField = <E as FieldElement>::BaseField>

The hash function used for hashing the rows of trace segment LDEs.

type VC: VectorCommitment<Self::HashFn>

The vector commitment scheme used for commiting to the trace.

Required Methods

fn get_main_trace_commitment(&self) -> <Self::HashFn as Hasher>::Digest

Returns the commitment to the low-degree extension of the main trace segment.

fn set_aux_trace( &mut self, aux_trace: &ColMatrix<E>, domain: &StarkDomain<<E as FieldElement>::BaseField>, ) -> (ColMatrix<E>, <Self::HashFn as Hasher>::Digest)

Takes auxiliary trace segment columns as input, interpolates them into polynomials in coefficient form, evaluates the polynomials over the LDE domain, and commits to the polynomial evaluations.

Returns a tuple containing the column polynomials in coefficient form and the commitment to the polynomial evaluations over the LDE domain.

Panics

This function is expected to panic if any of the following are true:

the number of rows in the provided aux_trace does not match the main trace.

this segment would exceed the number of segments specified by the trace layout.

fn read_main_trace_frame_into( &self, lde_step: usize, frame: &mut EvaluationFrame<<E as FieldElement>::BaseField>, )

Reads current and next rows from the main trace segment into the specified frame.

fn read_aux_trace_frame_into( &self, lde_step: usize, frame: &mut EvaluationFrame<E>, )

Reads current and next rows from the auxiliary trace segment into the specified frame.

fn read_lagrange_kernel_frame_into( &self, lde_step: usize, col_idx: usize, frame: &mut LagrangeKernelEvaluationFrame<E>, )

Populates the provided Lagrange kernel frame starting at the current row (as defined by lde_step).

Note that unlike EvaluationFrame, the Lagrange kernel frame includes only the Lagrange kernel column (as opposed to all columns).

fn query(&self, positions: &[usize]) -> Vec<Queries>

Returns trace table rows at the specified positions along with an opening proof to these rows.

fn trace_len(&self) -> usize

Returns the number of rows in the execution trace.

fn blowup(&self) -> usize

Returns blowup factor which was used to extend original execution trace into trace LDE.

fn trace_info(&self) -> &TraceInfo

Returns the trace info of the execution trace.